django/requirements.txt
django-rest-framework==0.1.0  
psycopg==3.2.3
langchain==0.1.16
langchain-openai==0.1.3
atlassian-python-api==3.41.16